pbd/pipelines/ocr_engine/pyproject.toml
[project]
name = "ocr-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bitsandbytes>=0.46.0",
    "boto3>=1.38.27",
    "kubernetes>=32.0.1",
    "minio>=7.2.15",
    "omegaconf>=2.3.0",
    "pillow>=11.2.1",
    "s3fs>=0.4.2",
    "vllm==0.9.0.1",
    "zenml==0.83.0",
]
